[arcane,std] Déplace 'ArcaneBasicVerifierService' dans son propre fichier.
